pyproject.toml
[tool.poetry]
name = "flaskapp"
version = "0.1.0"
description = "Basic Http Server"
authors = ["Safrizal <safrizal@kororo.co>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
